changes according to coding conventions
